requirements.txt
Django==4.0.6
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
PyJWT==2.4.0